BigInteger -> Integer

+++ src/main/java/com/architecture/rdb/common/AbstractCommonDAO.java
@@ -1,7 +1,6 @@
 package com.architecture.rdb.common;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
@@ -83,7 +82,7 @@ public abstract class AbstractCommonDAO<E extends CommonBean> implements CommonD
 	 * {@inheritDoc}.
 	 */
 	@TransactionAttribute(TransactionAttributeType.REQUIRES_NEW)
-	public void deleteById(BigInteger id) throws CommonException {
+	public void deleteById(Integer id) throws CommonException {
 		E obj = getEntityManager().find(getEntityClass(), id);
 		if (obj!= null){
 			this.delete(obj);
@@ -94,9 +93,9 @@ public abstract class AbstractCommonDAO<E extends CommonBean> implements CommonD
 	 * {@inheritDoc}.
 	 */
 	@TransactionAttribute(TransactionAttributeType.REQUIRES_NEW)
-	public void deleteByIds(List<BigInteger> listId) throws CommonException {
+	public void deleteByIds(List<Integer> listId) throws CommonException {
 		if (listId!= null && listId.size()>0){
-			for (BigInteger id : listId) {
+			for (Integer id : listId) {
 				this.deleteById(id);
 			}
 		}
@@ -107,7 +106,7 @@ public abstract class AbstractCommonDAO<E extends CommonBean> implements CommonD
 	 */
 	@SuppressWarnings("unchecked")
 	@TransactionAttribute(TransactionAttributeType.REQUIRED)
-	public List<E> findByIds(List<BigInteger> listId) throws CommonException {
+	public List<E> findByIds(List<Integer> listId) throws CommonException {
 		if (listId!= null && listId.size()>0){
 			try {
 				StringBuilder jpaQl = new StringBuilder();
@@ -291,7 +290,7 @@ public abstract class AbstractCommonDAO<E extends CommonBean> implements CommonD
 	 */
 	@SuppressWarnings("unchecked")
 	@TransactionAttribute(TransactionAttributeType.REQUIRED)
-	public E findById(BigInteger id) throws CommonException {
+	public E findById(Integer id) throws CommonException {
 		try{
 			StringBuilder jpaQl = new StringBuilder();
 			jpaQl.append(" SELECT entity FROM ");
